Implemented DatbaseException class

+++ pet/src/main/java/edu/alexandra/pet/PetModule/infrastructure/repository/PetRepositoryImpl.java
@@ -1,5 +1,6 @@
 package edu.alexandra.pet.PetModule.infrastructure.repository;
 
+import edu.alexandra.pet.PetModule.domain.exception.DatabaseException;
 import edu.alexandra.pet.PetModule.domain.model.Pet;
 import edu.alexandra.pet.PetModule.domain.port.PetRepositoryPort;
 import lombok.RequiredArgsConstructor;
@@ -16,6 +17,9 @@ public class PetRepositoryImpl implements PetRepositoryPort {
 
     @Override
     public void deletePet(String petId) {
+        if(!petMySQLRepository.existsById(petId)) {
+            throw new DatabaseException("Pet with ID " + petId + " not found");
+        }
         petMySQLRepository.deleteById(petId);
     }
 
@@ -31,18 +35,22 @@ public class PetRepositoryImpl implements PetRepositoryPort {
     public Pet getPet(String petId) {
         return petMySQLRepository.findById(petId)
                 .map(PetMapper::toDomain)
-                .orElseThrow(() -> new RuntimeException("Pet not found"));
+                .orElseThrow(() -> new DatabaseException("Pet not found"));
     }
 
     @Override
     public Pet createPet(Pet pet, String userId) {
-        return PetMapper.toDomain(petMySQLRepository.save(PetMapper.toEntity(pet, userId)));
+        try {
+            return PetMapper.toDomain(petMySQLRepository.save(PetMapper.toEntity(pet, userId)));
+        } catch (Exception e ) {
+            throw new DatabaseException("An unexpected error occurred while creating pet");
+        }
     }
 
     @Override
     public Pet updatePet(Pet pet) {
         PetEntity entity = petMySQLRepository.findById(pet.getId())
-                .orElseThrow(() -> new RuntimeException("Pet not found"));
+                .orElseThrow(() -> new DatabaseException("Pet not found"));
         return PetMapper.toDomain(petMySQLRepository.save(PetMapper.toEntity(pet, entity.getUserId())));
     }
 }
